requirements.txt
tomli~=2.0.1
docutils~=0.19
python-dotenv~=1.0.0
pytest~=7.4.0
qiskit~=0.43.3
pyyaml~=6.0.1
invoke~=2.2.0
flask~=2.3.2
alembic~=1.11.2
sqlalchemy~=2.0.19
click~=8.1.6
apispec~=6.3.0
marshmallow~=3.20.1
pyjwt~=2.8.0
werkzeug~=2.3.6
celery~=5.3.1
py-automapper~=1.2.3
